fix(reborn): tighten extension auth prompts
