+++ crates/ironclaw_auth/src/fakes.rs
@@ -24,10 +24,11 @@ use crate::{
     SecretSubmitResult, Timestamp, TurnGateAuthFlowQuery,
     cleanup::SecretCleanupAction::Deactivate,
     domain::{
-        PreparedCallbackFlow, account_is_authorized_for_requester, prepare_callback_flow,
-        recovery_projection_for_single_account, recovery_projection_for_unconfigured_accounts,
-        merge_provider_scopes, update_account_from_exchange, update_account_from_request,
-        validate_account_update_target, validate_bound_update_authority, validate_callback_claim,
+        PreparedCallbackFlow, account_is_authorized_for_requester, merge_provider_scopes,
+        prepare_callback_flow, recovery_projection_for_single_account,
+        recovery_projection_for_unconfigured_accounts, update_account_from_exchange,
+        update_account_from_request, validate_account_update_target,
+        validate_bound_update_authority, validate_callback_claim,
         validate_credential_status_transition, validate_flow_update_binding,
         validate_manual_token_flow, validate_manual_token_update_binding,
         validate_new_credential_account, validate_refresh_target, validate_selection_flow,
