+++ crates/ironclaw_first_party_extensions/src/gsuite/mod.rs
@@ -19,9 +19,8 @@ pub use handlers::{
     GsuiteDispatchRequest, GsuiteDispatchResult, GsuiteExecutor,
 };
 pub use manifest::{
-    CALENDAR_EXTENSION_ID, GMAIL_EXTENSION_ID, GSUITE_OUTPUT_BYTES_LIMIT,
-    GSUITE_PROVIDER_SCOPES, GSUITE_REQUEST_BODY_LIMIT, GSUITE_RESPONSE_BODY_LIMIT,
-    GSUITE_TIMEOUT_MS,
+    CALENDAR_EXTENSION_ID, GMAIL_EXTENSION_ID, GSUITE_OUTPUT_BYTES_LIMIT, GSUITE_PROVIDER_SCOPES,
+    GSUITE_REQUEST_BODY_LIMIT, GSUITE_RESPONSE_BODY_LIMIT, GSUITE_TIMEOUT_MS,
     GsuiteCapabilityOperation, GsuiteCapabilitySpec, GsuitePackageSpec, calendar_package_spec,
     find_gsuite_capability, gmail_package_spec, gsuite_package_specs, gsuite_resource_profile,
 };
