+++ crates/ironclaw_webui_v2_static/static/js/pages/extensions/hooks/useExtensions.js
@@ -223,7 +223,7 @@ export function useOauthSetup(packageRef) {
         popup.close();
       }
       refreshSetupState();
-      watchOauthProgress(authPopup);
+      if (authPopup) watchOauthProgress(authPopup);
     },
     onError: (_err, variables) => {
       const popup = variables?.popup;
